feat(API migration): Made some final changes
