+++ functions/model/tasks/editTask.js
@@ -6,7 +6,7 @@
 /* eslint-disable max-len */
 // eslint-disable-next-line no-dupe-else-if
 
-// const { db } = require("../application/lib");
+const { createSprintName } = require("../application/lib");
 const { addActivity } = require("../activity/addActivity");
 const { getOrgUseAppKey } = require("../organization/lib");
 const { getSprint, updateSprint, setSprint } = require("../sprints/lib");
@@ -126,13 +126,3 @@ exports.editTask = function(request, response) {
     });
     Promise.resolve(editTaskPromise);
 };
-
-function createSprintName(sprintId) {
-    if (sprintId == -1) {
-        return "Backlog";
-    } else if (sprintId == -2) {
-        return "Deleted";
-    } else {
-        return ("S" + sprintId.toString());
-    }
-}
