+++ functions/model/tasks/logwork.js
@@ -6,10 +6,10 @@
 /* eslint-disable max-len */
 // eslint-disable-next-line no-dupe-else-if
 
-// const { db } = require("../application/lib");
+const { createSprintName } = require("../application/lib");
 const { addActivity } = require("../activity/addActivity");
 const { getOrgUseAppKey } = require("../organization/lib");
-const { getOrgRawData, updateOrgRawData } = require("../orgRawData/lib");
+const { getOrgRawData, updateOrgRawData } = require("../organization/lib");
 const { getSprint, updateSprint } = require("../sprints/lib");
 const { getTask, updateTask } = require("./lib");
 
@@ -117,13 +117,3 @@ exports.logWork = function(request, response) {
     });
     return Promise.resolve(logWorkPromise);
 };
-
-function createSprintName(sprintId) {
-    if (sprintId == -1) {
-        return "Backlog";
-    } else if (sprintId == -2) {
-        return "Deleted";
-    } else {
-        return ("S" + sprintId.toString());
-    }
-}
